Replace Core Picture by UML Image
