+++ gaphor/UML/diagramitems.py
@@ -8,7 +8,7 @@ This module makes it easier to load a diagram item.
 from gaphor.UML.actions import *
 from gaphor.UML.classes import *
 from gaphor.UML.deployments import *
-from gaphor.UML.general import CommentItem, CommentLineItem
+from gaphor.UML.general import CommentItem, CommentLineItem, ImageItem
 from gaphor.UML.interactions import *
 from gaphor.UML.profiles import *
 from gaphor.UML.states import *
