+++ gaphor/UML/general/__init__.py
@@ -1,3 +1,4 @@
 from gaphor.UML.general import connectors as connectors
 from gaphor.UML.general.comment import CommentItem as CommentItem
 from gaphor.UML.general.comment import CommentLineItem as CommentLineItem
+from gaphor.UML.general.image import ImageItem as ImageItem
