+++ gaphor/UML/uicomponents.py
@@ -15,6 +15,6 @@ from gaphor.UML.classes import (
     dependencypropertypages,
     enumerationpropertypages,
 )
-from gaphor.UML.general import commenteditor, commentpropertypage
+from gaphor.UML.general import commenteditor, commentpropertypage, imagepropertypage
 from gaphor.UML.interactions import interactionspropertypages
 from gaphor.UML.profiles import metaclasspropertypage, stereotypepropertypages
