+++ gaphor/core/modeling/__init__.py
@@ -4,7 +4,6 @@ from gaphor.core.modeling.base import Base, Id, swap_element_type
 from gaphor.core.modeling.coremodel import (
     ElementChange,
     PendingChange,
-    Picture,
     RefChange,
     ValueChange,
 )
